+++ NuguAgents/Sources/CapabilityAgents/AutomaticSpeechRecognition/EndPointDetector/ClientEndPointDetector.swift
@@ -30,6 +30,7 @@ class ClientEndPointDetector: EndPointDetectable {
     private var engine: TycheEndPointDetectorEngine!
     
     private var boundStreams: AudioBoundStreams?
+    private let detectorQueue = DispatchQueue(label: "com.sktelecom.romaine.end_point_detector")
     
     private var state: EndPointDetectorState = .idle {
         didSet {
@@ -44,25 +45,39 @@ class ClientEndPointDetector: EndPointDetectable {
         engine.delegate = self
     }
     
+    deinit {
+        internalStop()
+    }
+    
     func start(audioStreamReader: AudioStreamReadable) {
         log.debug("start")
         
-        boundStreams?.stop()
-        let audioBoundStream = AudioBoundStreams(audioStreamReader: audioStreamReader)
-        boundStreams = audioBoundStream
+        detectorQueue.async { [weak self] in
+            guard let self = self else { return }
+            
+            self.internalStop()
+            self.boundStreams = AudioBoundStreams(audioStreamReader: audioStreamReader)
 
-        engine.start(
-            inputStream: audioBoundStream.input,
-            sampleRate: asrOptions.sampleRate,
-            timeout: asrOptions.timeout.truncatedSeconds,
-            maxDuration: asrOptions.maxDuration.truncatedSeconds,
-            pauseLength: asrOptions.pauseLength.truncatedMilliSeconds
-        )
+            self.engine.start(
+                inputStream: self.boundStreams!.input,
+                sampleRate: self.asrOptions.sampleRate,
+                timeout: self.asrOptions.timeout.truncatedSeconds,
+                maxDuration: self.asrOptions.maxDuration.truncatedSeconds,
+                pauseLength: self.asrOptions.pauseLength.truncatedMilliSeconds
+            )
+        }
+        
     }
     
     public func stop() {
         log.debug("stop")
         
+        detectorQueue.async { [weak self] in
+            self?.internalStop()
+        }
+    }
+    
+    private func internalStop() {
         boundStreams?.stop()
         engine.stop()
     }