+++ NuguAgents/Sources/Directive/PlayStackControl.swift
@@ -1,8 +1,8 @@
 //
-//  PlaySyncDelegate.swift
-//  NuguCore
+//  PlayStackControl.swift
+//  NuguAgents
 //
-//  Created by MinChul Lee on 2019/07/16.
+//  Created by MinChul Lee on 2020/07/10.
 //  Copyright (c) 2019 SK Telecom Co., Ltd. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,6 +20,21 @@
 
 import Foundation
 
-public protocol PlaySyncDelegate: class {
-    func playSyncDidRelease(property: PlaySyncProperty, messageId: String)
+public struct PlayStackControl {
+    let playServiceId: String?
+}
+
+
+// MARK: - Decodable
+
+extension PlayStackControl: Decodable {
+    enum CodingKeys: String, CodingKey {
+        case playServiceId
+    }
+    
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        
+        playServiceId = try? container.decode(String.self, forKey: .playServiceId)
+    }
 }