+++ NuguServiceKit/Sources/NuguServiceWebJavascriptDelegate.swift
@@ -23,4 +23,5 @@
 public protocol NuguServiceWebJavascriptDelegate: class {
     func openExternalApp(openExternalAppItem: WebOpenExternalApp)
     func openInAppBrowser(url: String)
+    func closeWindow()
 }