+++ NuguTests/NuguAgents/SystemAgentSpec.swift
@@ -42,7 +42,7 @@ class SystemAgentSpec: QuickSpec {
             describe("context") {
                 var contextInfo: ContextInfo?
                 
-                waitUntil(timeout: 0.05) { (done) in
+                waitUntil(timeout: 0.2) { (done) in
                     systemAgent.contextInfoRequestContext { (context) in
                         contextInfo = context
                         done()